+++ engine/inc/renderer/3d/pipeline/dynamic/core/bag/dynpip_color_bag.hpp
