+++ engine/inc/renderer/3d/pipeline/dynamic/dynamic_pipeline.hpp
@@ -12,7 +12,8 @@
 
 #include <tamtypes.h>
 #include "../renderer_3d_pipeline.hpp"
-#include "./dynpip_options.hpp"
+#include "./core/dynpip_options.hpp"
+#include "./core/dynpip_core.hpp"
 #include "renderer/core/renderer_core.hpp"
 #include "renderer/3d/mesh/dynamic/dynamic_mesh.hpp"
 
@@ -23,10 +24,16 @@ class DynamicPipeline : public Renderer3DPipeline {
   DynamicPipeline();
   ~DynamicPipeline();
 
+  DynPipCore core;
+
   void init(RendererCore* core);
 
   void onUse();
 
+  /**
+   * Render 3D via "meshes".
+   * This render() method is a bridge to core.render() method.
+   */
   void render(DynamicMesh* mesh, const DynPipOptions* options = nullptr);
 };
