+++ engine/inc/renderer/3d/pipeline/static/core/stapip_core.hpp
@@ -19,14 +19,14 @@
 
 namespace Tyra {
 
-class StaPipelineCore {
+class StaPipCore {
  public:
-  StaPipelineCore();
-  ~StaPipelineCore();
+  StaPipCore();
+  ~StaPipCore();
 
   void init(RendererCore* t_core);
 
-  /** Render 3D data via "bags" */
+  /** Render 3D via "bags" */
   void render(StaPipBag* data, StaPipBagPackagesBBox* bbox = nullptr);
 
   /** Get max vert count of VU1 qbuffer (for optimizations) */
