+++ engine/src/renderer/3d/pipeline/dynamic/core/bag/dynpip_color_bag.cpp
@@ -1,0 +1,24 @@
+/*
+# ______       ____   ___
+#   |     \/   ____| |___|
+#   |     |   |   \  |   |
+#-----------------------------------------------------------------------
+# Copyright 2022, tyra - https://github.com/h4570/tyra
+# Licenced under Apache License 2.0
+# Sandro Sobczyński <sandro.sobczynski@gmail.com>
+*/
+
+#include "renderer/3d/pipeline/dynamic/core/bag/dynpip_color_bag.hpp"
+
+namespace Tyra {
+
+DynPipColorBag::DynPipColorBag() {
+  singleFrom = nullptr;
+  singleTo = nullptr;
+  manyFrom = nullptr;
+  manyTo = nullptr;
+}
+
+DynPipColorBag::~DynPipColorBag() {}
+
+}  // namespace Tyra
