+++ engine/src/renderer/3d/pipeline/dynamic/core/bag/dynpip_lighting_bag.cpp
@@ -8,17 +8,17 @@
 # Sandro Sobczyński <sandro.sobczynski@gmail.com>
 */
 
-#include "renderer/3d/pipeline/dynamic/bag/dynpip_color_bag.hpp"
+#include "renderer/3d/pipeline/dynamic/core/bag/dynpip_lighting_bag.hpp"
 
 namespace Tyra {
 
-DynPipColorBag::DynPipColorBag() {
-  singleFrom = nullptr;
-  singleTo = nullptr;
-  manyFrom = nullptr;
-  manyTo = nullptr;
+DynPipLightingBag::DynPipLightingBag() {
+  lightMatrix = nullptr;
+  normalsFrom = nullptr;
+  normalsTo = nullptr;
+  dirLights = nullptr;
 }
 
-DynPipColorBag::~DynPipColorBag() {}
+DynPipLightingBag::~DynPipLightingBag() {}
 
 }  // namespace Tyra
