+++ engine/src/renderer/3d/pipeline/dynamic/core/programs/dynpip_c_vu1.vclpp
@@ -10,18 +10,18 @@
 ; Triangle list
 ; Cull = Standard PS2 way. clipw polys are culled.
 ; Animation
-; Directional lights
+; Colors
 ;---------------------------------------------------------------
 
 .syntax new
-.name DynPipVU1_D
+.name DynPipVU1_C
 .vu
 .init_vf_all
 .init_vi_all
 
 #include "src/renderer/3d/pipeline/shared/vcl_sml.i"
 #include "src/renderer/3d/pipeline/shared/tyra_macros.i"
-#include "inc/renderer/3d/pipeline/dynamic/programs/dynpip_vu1_shared_defines.h"
+#include "inc/renderer/3d/pipeline/dynamic/core/programs/dynpip_vu1_shared_defines.h"
 
 #define RGBA_STORE_OFFSET   0
 #define XYZ2_STORE_OFFSET   1
@@ -29,9 +29,9 @@
 --enter
 --endenter
 
-#vuprog DynPipVU1D
+#vuprog DynPipVU1C
 
- --cont
+--cont
 
 #endvuprog
