+++ engine/src/renderer/3d/pipeline/dynamic/core/programs/dynpip_tc_vu1.vclpp
@@ -10,18 +10,18 @@
 ; Triangle list
 ; Cull = Standard PS2 way. clipw polys are culled.
 ; Animation
-; Texture, directional lights
+; Lighting, texture, colors
 ;---------------------------------------------------------------
 
 .syntax new
-.name DynPipVU1_TD
+.name DynPipVU1_TC
 .vu
 .init_vf_all
 .init_vi_all
 
 #include "src/renderer/3d/pipeline/shared/vcl_sml.i"
 #include "src/renderer/3d/pipeline/shared/tyra_macros.i"
-#include "inc/renderer/3d/pipeline/dynamic/programs/dynpip_vu1_shared_defines.h"
+#include "inc/renderer/3d/pipeline/dynamic/core/programs/dynpip_vu1_shared_defines.h"
 
 #define STQ_STORE_OFFSET    0
 #define RGBA_STORE_OFFSET   1
@@ -30,7 +30,7 @@
 --enter
 --endenter
 
-#vuprog DynPipVU1TD
+#vuprog DynPipVU1TC
 
 --cont
