+++ engine/src/renderer/3d/pipeline/dynamic/core/programs/dynpip_tc_vu1_program.cpp
@@ -9,27 +9,27 @@
 */
 
 #include "debug/debug.hpp"
-#include "renderer/3d/pipeline/dynamic/programs/dynpip_td_vu1_program.hpp"
+#include "renderer/3d/pipeline/dynamic/core/programs/dynpip_tc_vu1_program.hpp"
 
-extern u32 DynPipVU1_TD_CodeStart __attribute__((section(".vudata")));
-extern u32 DynPipVU1_TD_CodeEnd __attribute__((section(".vudata")));
+extern u32 DynPipVU1_TC_CodeStart __attribute__((section(".vudata")));
+extern u32 DynPipVU1_TC_CodeEnd __attribute__((section(".vudata")));
 
 namespace Tyra {
 
-DynPipTDVU1Program::DynPipTDVU1Program()
-    : DynPipVU1Program(DynPipTextureDirLights, &DynPipVU1_TD_CodeStart,
-                       &DynPipVU1_TD_CodeEnd,
+DynPipTCVU1Program::DynPipTCVU1Program()
+    : DynPipVU1Program(DynPipTextureColor, &DynPipVU1_TC_CodeStart,
+                       &DynPipVU1_TC_CodeEnd,
                        ((u64)GIF_REG_ST) << 0 | ((u64)GIF_REG_RGBAQ) << 4 |
                            ((u64)GIF_REG_XYZ2) << 8,
-                       3, 4) {}
+                       3, 3) {}
 
-DynPipTDVU1Program::~DynPipTDVU1Program() {}
+DynPipTCVU1Program::~DynPipTCVU1Program() {}
 
-std::string DynPipTDVU1Program::getStringName() const {
-  return std::string("DynPip - TD");
+std::string DynPipTCVU1Program::getStringName() const {
+  return std::string("DynPip - TC");
 }
 
-void DynPipTDVU1Program::addProgramQBufferDataToPacket(
+void DynPipTCVU1Program::addProgramQBufferDataToPacket(
     packet2_t* packet, DynPipQBuffer* qbuffer) const {
   // u32 addr = VU1_VERT_DATA_ADDR;
 
@@ -41,11 +41,6 @@ void DynPipTDVU1Program::addProgramQBufferDataToPacket(
   // // Add sts
   // packet2_utils_vu_add_unpack_data(packet, addr, qbuffer->sts, qbuffer->size,
   //                                  true);
-  // addr += qbuffer->size;
-
-  // // Add normal
-  // packet2_utils_vu_add_unpack_data(packet, addr, qbuffer->normals,
-  //                                  qbuffer->size, true);
 
   // // Add colors
   // if (qbuffer->bag->color->single == nullptr) {
