+++ engine/src/renderer/3d/pipeline/dynamic/core/programs/dynpip_td_vu1.vclpp
@@ -10,26 +10,27 @@
 ; Triangle list
 ; Cull = Standard PS2 way. clipw polys are culled.
 ; Animation
-; Colors
+; Texture, directional lights
 ;---------------------------------------------------------------
 
 .syntax new
-.name DynPipVU1_C
+.name DynPipVU1_TD
 .vu
 .init_vf_all
 .init_vi_all
 
 #include "src/renderer/3d/pipeline/shared/vcl_sml.i"
 #include "src/renderer/3d/pipeline/shared/tyra_macros.i"
-#include "inc/renderer/3d/pipeline/dynamic/programs/dynpip_vu1_shared_defines.h"
+#include "inc/renderer/3d/pipeline/dynamic/core/programs/dynpip_vu1_shared_defines.h"
 
-#define RGBA_STORE_OFFSET   0
-#define XYZ2_STORE_OFFSET   1
+#define STQ_STORE_OFFSET    0
+#define RGBA_STORE_OFFSET   1
+#define XYZ2_STORE_OFFSET   2
 
 --enter
 --endenter
 
-#vuprog DynPipVU1C
+#vuprog DynPipVU1TD
 
 --cont
